Small refactor to simplify adding a GCP implementation
